+++ soo__loong/Makefile
@@ -9,15 +9,18 @@ LIBFT_DIR 	=	${SRC_DIR}/libft
 LIBFT		=	${SRC_DIR}/libft/libft.a
 
 SRCS 		=	${SRC_DIR}/main.c \
+				${UTILS_DIR}/animate.c \
 				${UTILS_DIR}/asset.c \
+				${UTILS_DIR}/data.c \
 				${UTILS_DIR}/draw.c \
 				${UTILS_DIR}/error.c \
 				${UTILS_DIR}/freer.c \
 				${UTILS_DIR}/game.c \
-				${UTILS_DIR}/map.c \
-				${UTILS_DIR}/validate_map.c \
 				${UTILS_DIR}/hook.c \
-				${UTILS_DIR}/tile.c
+				${UTILS_DIR}/map.c \
+				${UTILS_DIR}/tile.c \
+				${UTILS_DIR}/time.c \
+				${UTILS_DIR}/validate_map.c
 
 OBJS 		=	${SRCS:.c=.o}
 LIBFT_OBJS	=	$(wildcard $(LIBFT_DIR)/*.o)
